Refactor MOI-related tests into a single file and remove global state
