+++ test/conshdlr.jl
@@ -6,7 +6,6 @@
 @testset "dummy conshdlr (always satisfied, no constraint)" begin
     # create an empty problem
     o = SCIP.Optimizer()
-    MOI.set(o, MOI.Silent(), true)
     SCIP.set_parameter(o.inner, "display/verblevel", 0)
 
     # add the constraint handler
@@ -32,7 +31,7 @@ end
 @testset "dummy conshdlr (always satisfied, with constraint)" begin
     # create an empty problem
     o = SCIP.Optimizer()
-    MOI.set(o, MOI.Silent(), true)
+    SCIP.set_parameter(o.inner, "display/verblevel", 0)
 
     # add the constraint handler
     ch = Dummy.DummyConsHdlr()
@@ -60,7 +59,7 @@ end
 @testset "dummy conshdlr (always satisfied, no constraint, but needs it)" begin
     # create an empty problem
     o = SCIP.Optimizer()
-    MOI.set(o, MOI.Silent(), true)
+    SCIP.set_parameter(o.inner, "display/verblevel", 0)
 
     # add the constraint handler
     ch = Dummy.DummyConsHdlr()
@@ -85,7 +84,7 @@ end
 @testset "never satisfied conshdlr (does not need constraint)" begin
     # create an empty problem
     o = SCIP.Optimizer()
-    MOI.set(o, MOI.Silent(), true)
+    SCIP.set_parameter(o.inner, "display/verblevel", 0)
 
     # add the constraint handler
     ch = NeverSatisfied.NSCH()
@@ -110,7 +109,7 @@ end
 @testset "never satisfied conshdlr (needs constraint but does not have it)" begin
     # create an empty problem
     o = SCIP.Optimizer()
-    MOI.set(o, MOI.Silent(), true)
+    SCIP.set_parameter(o.inner, "display/verblevel", 0)
 
     # add the constraint handler
     ch = NeverSatisfied.NSCH()
@@ -135,7 +134,7 @@ end
 @testset "never satisfied conshdlr (needs constraint and has one)" begin
     # create an empty problem
     o = SCIP.Optimizer()
-    MOI.set(o, MOI.Silent(), true)
+    SCIP.set_parameter(o.inner, "display/verblevel", 0)
 
     # add the constraint handler
     ch = NeverSatisfied.NSCH()
